+++ .github/workflows/dependabot-auto-merge-go.yml
