+++ .github/workflows/lint.yml
@@ -13,4 +13,4 @@ jobs:
       - name: golangci-lint
         uses: golangci/golangci-lint-action@v3.1.0
         with:
-          version: latest
+          version: v1.42.1
