+++ go.mod
@@ -1,37 +1,34 @@
 module github.com/rustiever/conduit-connector-marketo
 
-go 1.18
+go 1.17
 
 require (
+	github.com/SpeakData/minimarketo v0.0.0-20170821092521-29339e452f44
 	github.com/conduitio/conduit-connector-sdk v0.2.0
 	github.com/goombaio/namegenerator v0.0.0-20181006234301-989e774b106e
+	github.com/jpillora/backoff v1.0.0
+	golang.org/x/sync v0.0.0-20220513210516-0976fa681c29
+	gopkg.in/tomb.v2 v2.0.0-20161208151619-d5d1b5820637
 )
 
 require (
 	github.com/conduitio/conduit-connector-protocol v0.2.0 // indirect
-	go.buf.build/library/go-grpc/conduitio/conduit-connector-protocol v1.4.1 // indirect
-	golang.org/x/net v0.0.0-20211108170745-6635138e15ea // indirect
-	google.golang.org/genproto v0.0.0-20211104193956-4c6863e31247 // indirect
-)
-
-require (
-	github.com/SpeakData/minimarketo v0.0.0-20170821092521-29339e452f44
-	github.com/fatih/color v1.13.0 // indirect
+	github.com/fatih/color v1.7.0 // indirect
 	github.com/golang/mock v1.6.0 // indirect
 	github.com/golang/protobuf v1.5.2 // indirect
-	github.com/hashicorp/go-hclog v1.2.0 // indirect
+	github.com/hashicorp/go-hclog v0.14.1 // indirect
 	github.com/hashicorp/go-plugin v1.4.3 // indirect
-	github.com/hashicorp/yamux v0.0.0-20211028200310-0bc27b27de87 // indirect
-	github.com/jpillora/backoff v1.0.0
-	github.com/mattn/go-colorable v0.1.12 // indirect
-	github.com/mattn/go-isatty v0.0.14 // indirect
-	github.com/mitchellh/go-testing-interface v1.14.1 // indirect
-	github.com/oklog/run v1.1.0 // indirect
+	github.com/hashicorp/yamux v0.0.0-20180604194846-3520598351bb // indirect
+	github.com/mattn/go-colorable v0.1.4 // indirect
+	github.com/mattn/go-isatty v0.0.10 // indirect
+	github.com/mitchellh/go-testing-interface v1.0.0 // indirect
+	github.com/oklog/run v1.0.0 // indirect
 	github.com/rs/zerolog v1.26.1 // indirect
-	golang.org/x/sync v0.0.0-20220513210516-0976fa681c29
-	golang.org/x/sys v0.0.0-20220519141025-dcacdad47464 // indirect
+	go.buf.build/library/go-grpc/conduitio/conduit-connector-protocol v1.4.1 // indirect
+	golang.org/x/net v0.0.0-20211108170745-6635138e15ea // indirect
+	golang.org/x/sys v0.0.0-20220405052023-b1e9470b6e64 // indirect
 	golang.org/x/text v0.3.7 // indirect
+	google.golang.org/genproto v0.0.0-20211104193956-4c6863e31247 // indirect
 	google.golang.org/grpc v1.43.0 // indirect
-	google.golang.org/protobuf v1.28.0 // indirect
-	gopkg.in/tomb.v2 v2.0.0-20161208151619-d5d1b5820637
+	google.golang.org/protobuf v1.27.1 // indirect
 )
