+++ source/iterator/combined_iterator.go
