+++ source/source_integration_test.go
@@ -11,33 +11,21 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package source
+package source_test
 
 import (
 	"context"
 	"encoding/json"
 	"errors"
-	"fmt"
-	"os"
 	"strings"
 	"testing"
 	"time"
 
-	"github.com/SpeakData/minimarketo"
 	sdk "github.com/conduitio/conduit-connector-sdk"
-	"github.com/goombaio/namegenerator"
 	"github.com/rustiever/conduit-connector-marketo/config"
-	marketoclient "github.com/rustiever/conduit-connector-marketo/marketo-client"
-	sourceConfig "github.com/rustiever/conduit-connector-marketo/source/config"
+	"github.com/rustiever/conduit-connector-marketo/source"
 	"github.com/rustiever/conduit-connector-marketo/source/iterator"
 	"github.com/rustiever/conduit-connector-marketo/source/position"
-)
-
-var (
-	ClinetID       = os.Getenv("MARKETO_CLIENT_ID")
-	ClientSecret   = os.Getenv("MARKETO_CLIENT_SECRET")
-	ClientEndpoint = os.Getenv("MARKETO_CLIENT_ENDPOINT")
-	fields         = []string{"firstName", "lastName", "email", "createdAt", "updatedAt"} // fields to be returned by the API
 )
 
 func TestSource_SuccessfullSnapshot(t *testing.T) {
@@ -47,7 +35,7 @@ func TestSource_SuccessfullSnapshot(t *testing.T) {
 		t.Fatal(err)
 	}
 	startTime := time.Now().UTC()
-	testLeads, err := AddLeads(client, 10)
+	testLeads, err := addLeads(client, 10)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -57,7 +45,7 @@ func TestSource_SuccessfullSnapshot(t *testing.T) {
 			t.Error(err)
 		}
 	})
-	src := &Source{}
+	src := &source.Source{}
 	ctx := context.Background()
 	defer func() {
 		_ = src.TearDown(ctx)
@@ -83,7 +71,7 @@ func TestSource_SnapshotRestart(t *testing.T) {
 		t.Fatal(err)
 	}
 	startTime := time.Now().UTC()
-	testLeads, err := AddLeads(client, 10)
+	testLeads, err := addLeads(client, 10)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -93,7 +81,7 @@ func TestSource_SnapshotRestart(t *testing.T) {
 			t.Error(err)
 		}
 	})
-	src := &Source{}
+	src := &source.Source{}
 	ctx := context.Background()
 	defer func() {
 		_ = src.TearDown(ctx)
@@ -123,7 +111,7 @@ func TestSource_SnapshotRestart(t *testing.T) {
 
 func TestSource_EmptyDatabase(t *testing.T) {
 	iterator.InitialDate = time.Now().UTC()
-	src := &Source{}
+	src := &source.Source{}
 	ctx := context.Background()
 	defer func() {
 		_ = src.TearDown(ctx)
@@ -141,7 +129,7 @@ func TestSource_EmptyDatabase(t *testing.T) {
 func TestSource_StartCDCAfterEmptyBucket(t *testing.T) {
 	iterator.InitialDate = time.Now().UTC()
 	ctx := context.Background()
-	src := &Source{}
+	src := &source.Source{}
 	defer func() {
 		_ = src.TearDown(ctx)
 	}()
@@ -166,7 +154,7 @@ func TestSource_StartCDCAfterEmptyBucket(t *testing.T) {
 			t.Error(err)
 		}
 	})
-	testLeads, err := AddLeads(client, 5)
+	testLeads, err := addLeads(client, 5)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -182,7 +170,7 @@ func TestSource_StartCDCAfterEmptyBucket(t *testing.T) {
 
 func TestSource_NonExistentDatabase(t *testing.T) {
 	iterator.InitialDate = time.Now().UTC()
-	src := &Source{}
+	src := &source.Source{}
 	ctx := context.Background()
 	defer func() {
 		_ = src.TearDown(ctx)
@@ -202,7 +190,7 @@ func TestSource_NonExistentDatabase(t *testing.T) {
 
 func TestSource_CDC_ReadRecordsUpdate(t *testing.T) {
 	iterator.InitialDate = time.Now().UTC()
-	src := &Source{}
+	src := &source.Source{}
 	ctx := context.Background()
 	defer func() {
 		_ = src.TearDown(ctx)
@@ -216,7 +204,7 @@ func TestSource_CDC_ReadRecordsUpdate(t *testing.T) {
 		t.Fatal(err)
 	}
 	startTime := time.Now().UTC()
-	testLeads, err := AddLeads(client, 1)
+	testLeads, err := addLeads(client, 1)
 	t.Cleanup(func() {
 		err := cleanUp(client, startTime)
 		if err != nil {
@@ -231,7 +219,7 @@ func TestSource_CDC_ReadRecordsUpdate(t *testing.T) {
 		t.Errorf("expected no error, got %v", err)
 	}
 	assert(t, &rec, testLeads[0])
-	updatedLeads, err := UpdateLeads(client, testLeads[0]["email"].(string))
+	updatedLeads, err := updateLeads(client, testLeads[0]["email"].(string))
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -249,7 +237,7 @@ func TestSource_CDC_ReadRecordsUpdate(t *testing.T) {
 func TestCDC_Delete(t *testing.T) {
 	iterator.InitialDate = time.Now().UTC()
 	ctx := context.Background()
-	src := &Source{}
+	src := &source.Source{}
 	defer func() {
 		_ = src.TearDown(ctx)
 	}()
@@ -268,7 +256,7 @@ func TestCDC_Delete(t *testing.T) {
 			t.Error(err)
 		}
 	})
-	testLeads, err := AddLeads(client, 1)
+	testLeads, err := addLeads(client, 1)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -287,7 +275,7 @@ func TestCDC_Delete(t *testing.T) {
 		t.Errorf("expected no error, got %v", err)
 	}
 	leadID := record["id"].(string)
-	err = client.DeleteLeadsByIDs([]string{leadID})
+	err = client.deleteLeadsByIDs([]string{leadID})
 	if err != nil {
 		t.Errorf("expected no error, got %v", err)
 	}
@@ -303,7 +291,7 @@ func TestCDC_Delete(t *testing.T) {
 
 func TestSource_CDC_ReadRecordsInsertAfterTeardown(t *testing.T) {
 	iterator.InitialDate = time.Now().UTC()
-	src := &Source{}
+	src := &source.Source{}
 	ctx := context.Background()
 	err := configAndOpen(ctx, src, nil)
 	if err != nil {
@@ -314,7 +302,7 @@ func TestSource_CDC_ReadRecordsInsertAfterTeardown(t *testing.T) {
 		t.Fatal(err)
 	}
 	startTime := time.Now().UTC()
-	testLeads, err := AddLeads(client, 3)
+	testLeads, err := addLeads(client, 3)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -331,7 +319,7 @@ func TestSource_CDC_ReadRecordsInsertAfterTeardown(t *testing.T) {
 	}
 	lastPosition := rec.Position
 	_ = src.TearDown(ctx)
-	src1 := &Source{}
+	src1 := &source.Source{}
 	defer func() {
 		_ = src1.TearDown(ctx)
 	}()
@@ -339,7 +327,7 @@ func TestSource_CDC_ReadRecordsInsertAfterTeardown(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	testLeads, err = AddLeads(client, 1)
+	testLeads, err = addLeads(client, 1)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -349,7 +337,7 @@ func TestSource_CDC_ReadRecordsInsertAfterTeardown(t *testing.T) {
 
 func TestOpenSource_FailsParsePosition(t *testing.T) {
 	ctx := context.Background()
-	source := &Source{}
+	source := &source.Source{}
 	defer func() {
 		_ = source.Teardown(ctx)
 	}()
@@ -366,7 +354,7 @@ func TestOpenSource_FailsParsePosition(t *testing.T) {
 
 func TestOpenSource_InvalidPositionType(t *testing.T) {
 	ctx := context.Background()
-	source := &Source{}
+	source := &source.Source{}
 	defer func() {
 		_ = source.Teardown(ctx)
 	}()
@@ -391,7 +379,7 @@ func TestOpenSource_InvalidPositionType(t *testing.T) {
 }
 
 func TestConfigureSource_FailsWhenConfigEmpty(t *testing.T) {
-	con := &Source{}
+	con := &source.Source{}
 	err := con.Configure(context.Background(), make(map[string]string))
 	if err == nil {
 		t.Errorf("expected no error, got %v", err)
@@ -399,141 +387,4 @@ func TestConfigureSource_FailsWhenConfigEmpty(t *testing.T) {
 	if strings.HasPrefix(err.Error(), "config is invalid:") {
 		t.Errorf("expected error to be about missing config, got %v", err)
 	}
-}
-
-// returns configs for testing.
-func getConfigs() map[string]string {
-	cfg := map[string]string{}
-	cfg[config.ClientID] = ClinetID
-	cfg[config.ClientSecret] = ClientSecret
-	cfg[config.ClientEndpoint] = ClientEndpoint
-	cfg[sourceConfig.ConfigKeyPollingPeriod] = "10s"
-	return cfg
-}
-
-// returns new client.
-func getClient() (marketoclient.Client, error) {
-	client, err := marketoclient.NewClient(minimarketo.ClientConfig{
-		ID:       ClinetID,
-		Secret:   ClientSecret,
-		Endpoint: ClientEndpoint,
-	})
-	if err != nil {
-		return marketoclient.Client{}, err
-	}
-	return client, nil
-}
-
-// asserts actual record against expected lead.
-func assert(t *testing.T, actual *sdk.Record, expected map[string]interface{}) {
-	var record map[string]interface{}
-	err := json.Unmarshal(actual.Payload.Bytes(), &record)
-	if err != nil {
-		t.Errorf("expected no error, got %v", err)
-	}
-	if expected["firstName"] != record["firstName"] {
-		t.Errorf("expected firstName %v, got %v", expected["firstName"], record["firstName"])
-	}
-	if expected["lastName"] != record["lastName"] {
-		t.Errorf("expected lastName %v, got %v", expected["lastName"], record["lastName"])
-	}
-	if expected["email"] != record["email"] {
-		t.Errorf("expected email %v, got %v", expected["email"], record["email"])
-	}
-}
-
-// generates a n number of leads and adds them to the database, also returns leads.
-func AddLeads(client marketoclient.Client, count int) ([]map[string]interface{}, error) {
-	seed := time.Now().UTC().UnixNano()
-	nameGenerator := namegenerator.NewNameGenerator(seed)
-	var leads []map[string]interface{}
-	for i := 0; i < count; i++ {
-		firstname := nameGenerator.Generate()
-		leads = append(leads, map[string]interface{}{
-			"firstName": firstname,
-			"lastName":  nameGenerator.Generate(),
-			"email":     firstname + "@meroxa.com",
-		})
-	}
-	err := client.CreateOrUpdateLeads(marketoclient.CreateOnly, leads)
-	if err != nil {
-		return nil, err
-	}
-	return leads, nil
-}
-
-// updates the leads for given LeadID
-func UpdateLeads(client marketoclient.Client, emailID string) (map[string]interface{}, error) {
-	seed := time.Now().UTC().UnixNano()
-	nameGenerator := namegenerator.NewNameGenerator(seed)
-	leads := map[string]interface{}{
-		"lastName": nameGenerator.Generate(),
-		"email":    emailID,
-	}
-	err := client.CreateOrUpdateLeads(marketoclient.UpdateOnly, []map[string]interface{}{leads})
-	if err != nil {
-		return nil, err
-	}
-	return leads, nil
-}
-
-// gets next record from the source
-func nextRecord(ctx context.Context, src *Source, t *testing.T) (rec sdk.Record) {
-	var err error
-	for {
-		rec, err = src.Read(ctx)
-		if errors.Is(err, sdk.ErrBackoffRetry) {
-			continue
-		}
-		if err != nil {
-			t.Errorf("expected no error, got %v", err)
-		}
-		break
-	}
-	return
-}
-
-// deletes all leads from marketo API
-func cleanUp(client marketoclient.Client, sinceTime time.Time) error {
-	token, err := client.GetNextPageToken(sinceTime)
-	if err != nil {
-		return err
-	}
-	res, err := client.GetLeadChanges(token, fields)
-	if err != nil {
-		return err
-	}
-	if len(res.Result) == 0 {
-		return nil
-	}
-	leadResult := []map[string]interface{}{}
-	err = json.Unmarshal(res.Result, &leadResult)
-	if err != nil {
-		return err
-	}
-	var leadIDs = make([]string, 0)
-	for _, data := range leadResult {
-		leadIDs = append(leadIDs, fmt.Sprint(data["leadId"].(float64)))
-	}
-	if len(leadIDs) == 0 {
-		return nil
-	}
-	err = client.DeleteLeadsByIDs(leadIDs)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// configures the source with the given configs and establishes a connection to Marketo
-func configAndOpen(ctx context.Context, s *Source, pos sdk.Position) error {
-	err := s.Configure(ctx, getConfigs())
-	if err != nil {
-		return err
-	}
-	err = s.Open(ctx, pos)
-	if err != nil {
-		return err
-	}
-	return nil
 }
